product-consumer-service/pyproject.toml
[project]
name = "product-consumer-service"
version = "1.0.0"
description = "Product consumer service for Pact workshop"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pact-python>=2.2.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

